Solution to Check Multiples

+++ Check Multiples@@6788dd63490637cfc44e1b9a/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    float a,b;
-    scanf("%f %f",&a,&b);
+    int a,b;
+    scanf("%d %d",&a,&b);
     if(a%b==0){
         printf("Yes");
     }
